Address review: fix metadata check consistency, move constant, add export test
- Fix inconsistent `if ingestion_metadata:` to `if ingestion_metadata is not None:`
  in _reconstruct_document_paths (import_tasks_v2.py) for consistency with the
  rest of the PR's truthiness-vs-None checks.

- Move EXPECTED_GLOBAL_ID_TYPE constant from ingestion_source_mutations.py to
  document_types.py to eliminate coupling from document_mutations.py and
  document_queries.py importing from a mutations module.

- Change export_v2.py ingestion_metadata check to truthiness (`if
  doc_path.ingestion_metadata:`) so empty-dict defaults are omitted from
  exports, avoiding noise.

- Add test_export_omits_empty_dict_ingestion_metadata test verifying that
  paths with the default empty-dict metadata do not include the field in
  export output.
