+++ config/graphql/document_types.py
@@ -46,6 +46,7 @@ logger = logging.getLogger(__name__)
 
 # -------------------- Ingestion Source Types -------------------- #
 
+EXPECTED_GLOBAL_ID_TYPE = "IngestionSourceType"
 
 IngestionSourceTypeEnum = graphene.Enum.from_enum(
     IngestionSourceCategory, name="IngestionSourceTypeEnum"
